Renamed to jupyter-keras-tract, cleaned Jupyter notebook, added shape optimization for Rust script
