+++ examples/jupyter-keras-tract/environment.yml
@@ -1,0 +1,108 @@
+name: tf
+channels:
+  - defaults
+dependencies:
+  - _libgcc_mutex=0.1=main
+  - _tflow_select=2.3.0=mkl
+  - absl-py=0.8.1=py37_0
+  - astor=0.8.0=py37_0
+  - attrs=19.3.0=py_0
+  - backcall=0.1.0=py37_0
+  - blas=1.0=mkl
+  - bleach=3.1.0=py37_0
+  - c-ares=1.15.0=h7b6447c_1001
+  - ca-certificates=2020.1.1=0
+  - certifi=2019.11.28=py37_0
+  - decorator=4.4.1=py_0
+  - defusedxml=0.6.0=py_0
+  - entrypoints=0.3=py37_0
+  - gast=0.2.2=py37_0
+  - gmp=6.1.2=h6c8ec71_1
+  - google-pasta=0.1.8=py_0
+  - grpcio=1.16.1=py37hf8bcb03_1
+  - h5py=2.10.0=py37h7918eee_0
+  - hdf5=1.10.4=hb1b8bf9_0
+  - importlib_metadata=1.5.0=py37_0
+  - inflect=4.1.0=py37_0
+  - intel-openmp=2020.0=166
+  - ipykernel=5.1.4=py37h39e3cac_0
+  - ipython=7.12.0=py37h5ca1d4c_0
+  - ipython_genutils=0.2.0=py37_0
+  - jaraco.itertools=5.0.0=py_0
+  - jedi=0.16.0=py37_0
+  - jinja2=2.11.1=py_0
+  - jsonschema=3.2.0=py37_0
+  - jupyter_client=5.3.4=py37_0
+  - jupyter_core=4.6.1=py37_0
+  - keras-applications=1.0.8=py_0
+  - keras-preprocessing=1.1.0=py_1
+  - ld_impl_linux-64=2.33.1=h53a641e_7
+  - libedit=3.1.20181209=hc058e9b_0
+  - libffi=3.2.1=hd88cf55_4
+  - libgcc-ng=9.1.0=hdf63c60_0
+  - libgfortran-ng=7.3.0=hdf63c60_0
+  - libprotobuf=3.11.2=hd408876_0
+  - libsodium=1.0.16=h1bed415_0
+  - libstdcxx-ng=9.1.0=hdf63c60_0
+  - markdown=3.1.1=py37_0
+  - markupsafe=1.1.1=py37h7b6447c_0
+  - mistune=0.8.4=py37h7b6447c_0
+  - mkl=2020.0=166
+  - mkl-service=2.3.0=py37he904b0f_0
+  - mkl_fft=1.0.15=py37ha843d7b_0
+  - mkl_random=1.1.0=py37hd6b4f25_0
+  - more-itertools=8.2.0=py_0
+  - nb_conda=2.2.1=py37_0
+  - nb_conda_kernels=2.2.2=py37_0
+  - nbconvert=5.6.1=py37_0
+  - nbformat=5.0.4=py_0
+  - ncurses=6.1=he6710b0_1
+  - notebook=6.0.3=py37_0
+  - numpy=1.18.1=py37h4f9e942_0
+  - numpy-base=1.18.1=py37hde5b4d6_1
+  - openssl=1.1.1d=h7b6447c_3
+  - opt_einsum=3.1.0=py_0
+  - pandas=1.0.0=py37h0573a6f_0
+  - pandoc=2.2.3.2=0
+  - pandocfilters=1.4.2=py37_1
+  - parso=0.6.0=py_0
+  - pexpect=4.8.0=py37_0
+  - pickleshare=0.7.5=py37_0
+  - pip=20.0.2=py37_1
+  - prometheus_client=0.7.1=py_0
+  - prompt_toolkit=3.0.3=py_0
+  - protobuf=3.11.2=py37he6710b0_0
+  - ptyprocess=0.6.0=py37_0
+  - pygments=2.5.2=py_0
+  - pyrsistent=0.15.7=py37h7b6447c_0
+  - python=3.7.6=h0371630_2
+  - python-dateutil=2.8.1=py_0
+  - pytz=2019.3=py_0
+  - pyzmq=18.1.1=py37he6710b0_0
+  - readline=7.0=h7b6447c_5
+  - scipy=1.4.1=py37h0b6359f_0
+  - send2trash=1.5.0=py37_0
+  - setuptools=45.1.0=py37_0
+  - six=1.14.0=py37_0
+  - sqlite=3.31.1=h7b6447c_0
+  - tensorboard=2.0.0=pyhb38c66f_1
+  - tensorflow=2.0.0=mkl_py37h66b46cc_0
+  - tensorflow-base=2.0.0=mkl_py37h9204916_0
+  - tensorflow-estimator=2.0.0=pyh2649769_0
+  - termcolor=1.1.0=py37_1
+  - terminado=0.8.3=py37_0
+  - testpath=0.4.4=py_0
+  - tk=8.6.8=hbc83047_0
+  - tornado=6.0.3=py37h7b6447c_1
+  - traitlets=4.3.3=py37_0
+  - wcwidth=0.1.8=py_0
+  - webencodings=0.5.1=py37_1
+  - werkzeug=0.16.1=py_0
+  - wheel=0.34.2=py37_0
+  - wrapt=1.11.2=py37h7b6447c_0
+  - xz=5.2.4=h14c3975_4
+  - zeromq=4.3.1=he6710b0_3
+  - zipp=2.1.0=py_0
+  - zlib=1.2.11=h7b6447c_3
+prefix: /home/matt/anaconda3/envs/tf
+
